re organized packages in initiative
